requirements.openvoice.txt
# This file contains the specific dependencies for the OpenVoice feature.
# It includes versions that may conflict with the main project, so it's used in a separate Docker build.

# --- Base Project Requirements ---
nltk==3.9.1
parler_tts @ git+https://github.com/huggingface/parler-tts.git
melotts @ git+https://github.com/andimarafioti/MeloTTS.git#egg=MeloTTS
torch>=2.4.0
sounddevice==0.5.0
ChatTTS>=0.1.1
funasr>=1.1.6
modelscope>=1.17.1
deepfilternet>=0.5.6
openai>=1.40.1
useful-moonshine @ git+https://github.com/andimarafioti/moonshine.git

# --- OpenVoice V2 Dependencies (with specific versions) ---
faster-whisper>=1.0.3 # Pinned to match the main project's version
librosa==0.9.1
pydub==0.25.1
wavmark==0.0.3
eng_to_ipa==0.0.2
inflect==7.0.0
unidecode==1.3.7
whisper-timestamped==1.14.2
python-dotenv
pypinyin==0.50.0
cn2an==0.5.22
jieba==0.42.1
gradio==3.48.0
langid==1.1.6

# Install OpenVoice package directly from GitHub under its correct name
MyShell-OpenVoice @ git+https://github.com/lonelygo/OpenVoice.git
